Update InsertNodeRecursive.cpp
The question required to insert a node in LL recursively. I'm proposing you to kindly update your code, as the updated code fulfill the question requirement and executes the code correctly.
Description of Functions:
print(...) : It traverses through the linked list and prints it.

input(...) : It takes input the elements of linked list until -1 is met, and returns the head address.

insert(...):It inserts the given data at the given position using RECURSION.

Thank You.